+++ src/index.ts
@@ -1,2 +1,3 @@
 export { query } from "./core/query";
+export { arrayPipeline } from "./core/array-pipeline";
 export { getPathSegmentAfter } from "./helpers/path";